moved the test sensor option to test base material
